# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Universe Creator Platform API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
  message?: string;
}

export interface RegisterRequest {
  /** @minLength 3 */
  username: string;
  email: string;
  /** @minLength 8 */
  password: string;
  displayName?: string;
}

export interface LoginRequest {
  email: string;
  password: string;
}

export interface UserProfile {
  id: number;
  username: string;
  email: string;
  displayName?: string | null;
  avatarUrl?: string | null;
  bio?: string | null;
  createdAt: string;
}

export interface AuthResponse {
  token: string;
  user: UserProfile;
}

export type ProjectStatus = typeof ProjectStatus[keyof typeof ProjectStatus];


export const ProjectStatus = {
  draft: 'draft',
  published: 'published',
  archived: 'archived',
} as const;

export type ProjectMetadata = { [key: string]: unknown };

export interface Project {
  id: number;
  userId: number;
  name: string;
  slug: string;
  description?: string | null;
  thumbnailUrl?: string | null;
  status: ProjectStatus;
  tags: string[];
  metadata: ProjectMetadata;
  createdAt: string;
  updatedAt: string;
}

export interface ProjectListResponse {
  items: Project[];
  total: number;
  limit: number;
  offset: number;
}

export interface CreateProjectRequest {
  name: string;
  description?: string;
  tags?: string[];
  templateId?: number;
}

export type UpdateProjectRequestStatus = typeof UpdateProjectRequestStatus[keyof typeof UpdateProjectRequestStatus];


export const UpdateProjectRequestStatus = {
  draft: 'draft',
  published: 'published',
  archived: 'archived',
} as const;

export type UpdateProjectRequestMetadata = { [key: string]: unknown };

export interface UpdateProjectRequest {
  name?: string;
  description?: string;
  status?: UpdateProjectRequestStatus;
  tags?: string[];
  thumbnailUrl?: string;
  metadata?: UpdateProjectRequestMetadata;
}

export type AssetType = typeof AssetType[keyof typeof AssetType];


export const AssetType = {
  image: 'image',
  audio: 'audio',
  video: 'video',
  model: 'model',
  document: 'document',
  other: 'other',
} as const;

export type AssetMetadata = { [key: string]: unknown };

export interface Asset {
  id: number;
  userId: number;
  projectId?: number | null;
  type: AssetType;
  name: string;
  filename: string;
  url: string;
  size?: number | null;
  mimeType?: string | null;
  metadata: AssetMetadata;
  createdAt: string;
}

export interface AssetListResponse {
  items: Asset[];
  total: number;
  limit: number;
  offset: number;
}

export type CreateAssetRequestType = typeof CreateAssetRequestType[keyof typeof CreateAssetRequestType];


export const CreateAssetRequestType = {
  image: 'image',
  audio: 'audio',
  video: 'video',
  model: 'model',
  document: 'document',
  other: 'other',
} as const;

export type CreateAssetRequestMetadata = { [key: string]: unknown };

export interface CreateAssetRequest {
  name: string;
  filename: string;
  url: string;
  type: CreateAssetRequestType;
  projectId?: number;
  size?: number;
  mimeType?: string;
  metadata?: CreateAssetRequestMetadata;
}

export interface Template {
  id: number;
  name: string;
  slug: string;
  description?: string | null;
  thumbnailUrl?: string | null;
  category?: string | null;
  tags: string[];
  usageCount: number;
  createdAt: string;
}

export interface TemplateListResponse {
  items: Template[];
  total: number;
  limit: number;
  offset: number;
}

export interface UseTemplateRequest {
  name: string;
}

export interface Plugin {
  id: number;
  name: string;
  slug: string;
  version: string;
  description?: string | null;
  isEnabled: boolean;
  createdAt: string;
}

export interface PluginListResponse {
  items: Plugin[];
  total: number;
  limit: number;
  offset: number;
}

export interface Package {
  id: number;
  projectId: number;
  name: string;
  version: string;
  description?: string | null;
  downloadUrl?: string | null;
  createdAt: string;
}

export interface PackageListResponse {
  items: Package[];
  total: number;
  limit: number;
  offset: number;
}

export type PublishRequestPayload = { [key: string]: unknown };

export interface PublishRequest {
  projectId: number;
  payload?: PublishRequestPayload;
}

export type PublishJobStatus = typeof PublishJobStatus[keyof typeof PublishJobStatus];


export const PublishJobStatus = {
  pending: 'pending',
  processing: 'processing',
  success: 'success',
  failed: 'failed',
} as const;

export interface PublishJob {
  id: number;
  projectId: number;
  userId: number;
  status: PublishJobStatus;
  errorMessage?: string | null;
  createdAt: string;
  completedAt?: string | null;
}

export interface ActivityItem {
  id: number;
  action: string;
  message?: string | null;
  createdAt: string;
}

export interface DashboardResponse {
  totalProjects: number;
  publishedProjects: number;
  totalAssets: number;
  totalDocuments: number;
  recentProjects: Project[];
  recentActivity: ActivityItem[];
}

export type ListProjectsParams = {
status?: ListProjectsStatus;
limit?: number;
offset?: number;
};

export type ListProjectsStatus = typeof ListProjectsStatus[keyof typeof ListProjectsStatus];


export const ListProjectsStatus = {
  draft: 'draft',
  published: 'published',
  archived: 'archived',
} as const;

export type ListAssetsParams = {
projectId?: number;
type?: ListAssetsType;
limit?: number;
offset?: number;
};

export type ListAssetsType = typeof ListAssetsType[keyof typeof ListAssetsType];


export const ListAssetsType = {
  image: 'image',
  audio: 'audio',
  video: 'video',
  model: 'model',
  document: 'document',
  other: 'other',
} as const;

export type ListTemplatesParams = {
category?: string;
limit?: number;
offset?: number;
};

export type ListPluginsParams = {
limit?: number;
offset?: number;
};

export type ListPackagesParams = {
projectId?: number;
limit?: number;
offset?: number;
};