# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Universe Creator Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: authRegister
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: User already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: authLogin
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/me:
    get:
      operationId: authMe
      tags: [auth]
      summary: Get current user
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /assets:
    get:
      operationId: listAssets
      tags: [assets]
      summary: List assets
      security:
        - bearerAuth: []
      parameters:
        - name: projectId
          in: query
          schema:
            type: integer
        - name: type
          in: query
          schema:
            type: string
            enum: [image, audio, video, model, document, other]
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: List of assets
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AssetListResponse"
    post:
      operationId: createAsset
      tags: [assets]
      summary: Upload/register an asset
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAssetRequest"
      responses:
        "201":
          description: Asset created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Asset"
  /assets/{id}:
    get:
      operationId: getAsset
      tags: [assets]
      summary: Get asset by ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Asset
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Asset"
    delete:
      operationId: deleteAsset
      tags: [assets]
      summary: Delete asset
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /plugins:
    get:
      operationId: listPlugins
      tags: [plugins]
      summary: List plugins
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: List of plugins
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PluginListResponse"
  /plugins/{id}:
    get:
      operationId: getPlugin
      tags: [plugins]
      summary: Get plugin by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Plugin
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Plugin"
  /packages:
    get:
      operationId: listPackages
      tags: [packages]
      summary: List packages
      security:
        - bearerAuth: []
      parameters:
        - name: projectId
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: List of packages
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PackageListResponse"
  /packages/{id}:
    get:
      operationId: getPackage
      tags: [packages]
      summary: Get package by ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Package
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Package"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
    RegisterRequest:
      type: object
      properties:
        username:
          type: string
          minLength: 3
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
        displayName:
          type: string
      required:
        - username
        - email
        - password
    LoginRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
      required:
        - email
        - password
    AuthResponse:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/UserProfile"
      required:
        - token
        - user
    UserProfile:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        email:
          type: string
        displayName:
          type: string
          nullable: true
        avatarUrl:
          type: string
          nullable: true
        bio:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - username
        - email
        - createdAt
...
    Asset:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        projectId:
          type: integer
          nullable: true
        type:
          type: string
          enum: [image, audio, video, model, document, other]
        name:
          type: string
        filename:
          type: string
        url:
          type: string
        size:
          type: integer
          nullable: true
        mimeType:
          type: string
          nullable: true
        metadata:
          type: object
          additionalProperties: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - type
        - name
        - filename
        - url
        - metadata
        - createdAt
    AssetListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Asset"
        total:
          type: integer
        limit:
          type: integer
        offset:
          type: integer
      required:
        - items
        - total
        - limit
        - offset
    CreateAssetRequest:
      type: object
      properties:
        name:
          type: string
        filename:
          type: string
        url:
          type: string
        type:
          type: string
          enum: [image, audio, video, model, document, other]
        projectId:
          type: integer
        size:
          type: integer
        mimeType:
          type: string
        metadata:
          type: object
          additionalProperties: true
      required:
        - name
        - filename
        - url
        - type
...
    Plugin:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        version:
          type: string
        description:
          type: string
          nullable: true
        isEnabled:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - slug
        - version
        - isEnabled
        - createdAt
    PluginListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Plugin"
        total:
          type: integer
        limit:
          type: integer
        offset:
          type: integer
      required:
        - items
        - total
        - limit
        - offset
    Package:
      type: object
      properties:
        id:
          type: integer
        projectId:
          type: integer
        name:
          type: string
        version:
          type: string
        description:
          type: string
          nullable: true
        downloadUrl:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - projectId
        - name
        - version
        - createdAt
    PackageListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Package"
        total:
          type: integer
        limit:
          type: integer
        offset:
          type: integer
      required:
        - items
        - total
        - limit
        - offset
...