# lib/api-zod/src/generated/types/asset.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Universe Creator Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { AssetMetadata } from './assetMetadata';
import type { AssetType } from './assetType';

export interface Asset {
  id: number;
  userId: number;
  projectId?: number | null;
  type: AssetType;
  name: string;
  filename: string;
  url: string;
  size?: number | null;
  mimeType?: string | null;
  metadata: AssetMetadata;
  createdAt: Date;
}